Added functionality to add data and added finishing touches

+++ datavisual/dashboard/models.py
@@ -3,7 +3,7 @@ from django.db import models
 # Create your models here.
 class CountryData(models.Model):
     country = models.CharField(max_length=100)
-    population = models.IntegerField()
+    population = models.DecimalField(max_digits=10, decimal_places=4)
 
     class Meta:
         verbose_name_plural = 'Country Population Data'
